Add hitscan damager
Adds hitscan damager which deals damage based on a starting point,
endpoint and range. Saves origin and endpoint for e.g. latter
animation.

Adds get_cooldown_fraction to Dmgr_instance, this could be useful
for e.g. display of cooldown (or simple displaying of tracers).

Adds drawColorLine to window which allows to display a line
based on start and end point.